requirements.txt
blinker==1.9.0
click==8.2.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
pillow==11.2.1
setuptools==80.9.0
Werkzeug==3.1.3
